sql-py/requirements.txt
greenlet==2.0.2
inflect==6.0.2
pydantic==1.10.6
PyMySQL==1.0.2
python-dotenv==1.0.0
sqlacodegen==2.3.0.post1
SQLAlchemy==1.4.46
typing_extensions==4.5.0
